Rename and add js test
