+++ src/shared/PreApprovalRequest/PreApprovalPanel.test.js
@@ -1,9 +1,7 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 import { Provider } from 'react-redux';
-
 import configureStore from 'redux-mock-store';
-import { mount } from 'enzyme';
 
 import { PreApprovalPanel } from './PreApprovalPanel';
